pyproject.toml
[project]
name = "huggingface-image-dataset-generator"
version = "0.1.0"
description = "huggingface image dataset generator"
authors = [
    { name = "yashikota", email = "52403688+yashikota@users.noreply.github.com" }
]
dependencies = [
    "datasets>=2.19.1",
    "pillow>=10.3.0",
    "tqdm>=4.66.4",
    "huggingface_hub>=0.23.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

